requirements.txt
flask==2.0.1
python-dotenv==0.19.0
azure-servicebus==7.8.1
openai==1.0.0
werkzeug==2.0.1